# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MongoVision API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections/{connectionId}/databases/{dbName}/collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all collections in a database
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Collection"
                required: [collections]
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a new collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "200":
          description: Collection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}:
    delete:
      operationId: dropCollection
      tags: [collections]
      summary: Drop a collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection dropped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents in a collection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: filter
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: projection
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentsPage"
    post:
      operationId: insertDocument
      tags: [documents]
      summary: Insert a document
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertDocumentBody"
      responses:
        "200":
          description: Document inserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsertResult"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/documents/{documentId}:
    put:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResult"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /connections/{connectionId}/databases/{dbName}/collections/{collectionName}/bulk:
    post:
      operationId: bulkOperation
      tags: [documents]
      summary: Perform bulk operations on documents
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: dbName
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkOperationBody"
      responses:
        "200":
          description: Bulk operation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkOperationResult"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Collection:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        documentCount:
          type: number
        storageSize:
          type: number
        totalIndexSize:
          type: number
        avgDocumentSize:
          type: number
      required: [name, type]
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
      required: [name]
    DocumentsPage:
      type: object
      properties:
        documents:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
        executionTimeMs:
          type: number
      required: [documents, total, page, limit, totalPages]
    InsertDocumentBody:
      type: object
      properties:
        document:
          type: object
          additionalProperties: true
      required: [document]
    InsertResult:
      type: object
      properties:
        insertedId:
          type: string
        acknowledged:
          type: boolean
      required: [insertedId, acknowledged]
    UpdateDocumentBody:
      type: object
      properties:
        update:
          type: object
          additionalProperties: true
        replace:
          type: boolean
      required: [update]
    UpdateResult:
      type: object
      properties:
        matchedCount:
          type: number
        modifiedCount:
          type: number
        acknowledged:
          type: boolean
      required: [matchedCount, modifiedCount, acknowledged]
    BulkOperationBody:
      type: object
      properties:
        operation:
          type: string
          enum: [deleteMany, updateMany, insertMany]
        filter:
          type: object
          additionalProperties: true
        update:
          type: object
          additionalProperties: true
        documents:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [operation]
    BulkOperationResult:
      type: object
      properties:
        matchedCount:
          type: number
        modifiedCount:
          type: number
        deletedCount:
          type: number
        insertedCount:
          type: number
        acknowledged:
          type: boolean
      required: [acknowledged]
...