# lib/api-zod/src/generated/types/insertResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MongoVision API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InsertResult {
  insertedId: string;
  acknowledged: boolean;
}